split out IcebergUtil from S3DataLakeutil
