+++ airbyte-integrations/connectors/destination-s3-data-lake/src/test-integration/kotlin/io/airbyte/integrations/destination/s3_data_lake/S3DataLakeTestUtil.kt
@@ -10,9 +10,11 @@ import io.airbyte.cdk.command.ValidatedJsonUtils
 import io.airbyte.cdk.load.command.aws.AwsAssumeRoleCredentials
 import io.airbyte.cdk.load.command.aws.AwsEnvVarConstants
 import io.airbyte.cdk.load.util.Jsons
+import io.airbyte.integrations.destination.s3_data_lake.io.IcebergUtil
 import io.airbyte.integrations.destination.s3_data_lake.io.S3DataLakeUtil
 import java.nio.file.Files
 import java.nio.file.Path
+import org.apache.iceberg.catalog.Catalog
 
 object S3DataLakeTestUtil {
     val GLUE_CONFIG_PATH: Path = Path.of("secrets/glue.json")
@@ -47,9 +49,10 @@ object S3DataLakeTestUtil {
     fun getCatalog(
         config: S3DataLakeConfiguration,
         awsAssumeRoleCredentials: AwsAssumeRoleCredentials
-    ) =
-        S3DataLakeUtil(SimpleTableIdGenerator(), awsAssumeRoleCredentials).let { icebergUtil ->
-            val props = icebergUtil.toCatalogProperties(config)
-            icebergUtil.createCatalog(DEFAULT_CATALOG_NAME, props)
-        }
+    ): Catalog {
+        val icebergUtil = IcebergUtil(SimpleTableIdGenerator())
+        val s3DataLakeUtil = S3DataLakeUtil(icebergUtil, awsAssumeRoleCredentials)
+        val props = s3DataLakeUtil.toCatalogProperties(config)
+        return icebergUtil.createCatalog(DEFAULT_CATALOG_NAME, props)
+    }
 }
